api/requirements.txt
cached-property==1.3.0
click==6.7
colorama==0.3.7
configobj==5.0.6
Django==1.10.5
djangorestframework==3.5.3
docker==2.0.2
docker-compose==1.10.0
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
environs==1.2.0
humanize==0.5.1
jsonschema==2.5.1
marshmallow==2.12.1
pgcli==1.4.0
pgspecial==1.7.0
prompt-toolkit==1.0.9
psycopg2==2.6.2
Pygments==2.2.0
PyYAML==3.12
read-env==1.1.0
requests==2.11.1
setproctitle==1.1.10
six==1.10.0
sqlparse==0.2.2
texttable==0.8.7
wcwidth==0.1.7
websocket-client==0.40.0
